requirements.txt
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.23
django-cors-headers==4.7.0
django_celery_results==2.6.0
djangorestframework==3.16.0
gunicorn==23.0.0
kombu==5.5.4
packaging==25.0
pathspec==0.11.2
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
